Use rb_obj_id() as the key in rb2js hash.
Hashing the real objects seems to make for unhappiness, mostly to do
with the amount of interaction involved upon calling rb_hash_delete,
which occurs in finalize, which is called from SpiderMonkey GC... which
is really not a good time to be trying to call arbitrary methods back in
JS. So, let's not do that.

Also, const-ified a variable to silence a warning.